# metal : support MTLGPUFamily < Apple7, formatting, style