@jhen0409 2023-10-07
M2 (10c GPU), it looks like the speed has some slow down in some cases:

| model          |       size | backend     | ngl | test   | master t/s        | PR t/s            | speedup  |
| -------------- | ---------: | ----------- | --: | ------ | ----------------: | ----------------: | -------: |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 1   | 19.80 ± 5.49      | 17.70 ± 9.79      | 0.8944   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 2   | 36.79 ± 0.18      | 36.66 ± 0.05      | 0.9965   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 4   | 21.54 ± 0.03      | 36.77 ± 0.06      | 1.7079   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 5   | 26.76 ± 0.02      | 36.96 ± 0.16      | 1.3809   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 6   | 32.03 ± 0.07      | 39.96 ± 0.04      | 1.2483   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 7   | 37.26 ± 0.07      | 40.11 ± 0.07      | 1.0768   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 8   | 42.55 ± 0.04      | 38.42 ± 0.03      | 0.9030   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 9   | 47.65 ± 0.10      | 41.45 ± 0.05      | 0.8698   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 10  | 52.77 ± 0.08      | 41.57 ± 0.05      | 0.7878   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 11  | 57.77 ± 0.08      | 38.75 ± 0.13      | 0.6709   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 12  | 62.84 ± 0.06      | 39.08 ± 0.20      | 0.6218   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 13  | 67.62 ± 0.11      | 39.23 ± 0.09      | 0.5802   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 14  | 72.84 ± 0.04      | 41.79 ± 0.07      | 0.5739   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 15  | 77.63 ± 0.14      | 41.85 ± 0.05      | 0.5394   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 16  | 82.73 ± 0.17      | 82.65 ± 0.22      | 0.9990   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 32  | 166.89 ± 0.32     | 166.90 ± 0.31     | 1.0001   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 64  | 182.12 ± 0.28     | 182.21 ± 0.17     | 1.0005   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 128 | 185.47 ± 0.08     | 185.67 ± 0.17     | 1.0011   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | pp 256 | 183.72 ± 0.19     | 183.76 ± 0.19     | 1.0002   |
| llama 7B Q4_0  |   3.56 GiB |  Metal      |   1 | tg 128 | 21.92 ± 0.01      | 21.89 ± 0.01      | 0.9986   |

build: 99ed03a (1343)