@ggerganov 2023-10-07
Yes, I've also realized that the break-even point is not so trivial as currently proposed. It is function of the matrix sizes.
Still trying to find a general way to determine it