@ggerganov 2023-10-07
Bummer. I can't figure out an universal way to determine which kernel to use when. The break-even point for the number of batches at which the matrix-matrix kernel becomes more performant than the matrix-vector kernel depends both on the specifics of the hardware which are not queryable (number of cores, memory bandwidth, FLOPs) and on the model / matrix sizes.

Based on the tests here, there is a significant performance to be gained for quantized low-batch (< 16) decoding, which would be quite important for speculative approaches. But can't figure out a way to choose the optimal kernel. Any suggestions?