@KerfuffleV2 2023-10-07
> although similar analysis should be performed on the CUDA kernels

Are there actually already two kernels for CUDA? I actually wanted to mess with this and see if it helped with my issue where parallel generation gets slower and slower but it didn't look like there was that kind of logic in `ggml-cuda.cu`.

> But can't figure out a way to choose the optimal kernel. Any suggestions?

You didn't say "good suggestions". The simplest thing that comes to mind is to just do a mini benchmark that runs a few operations and stores the result into the context. Maybe that could even be part of the warmup stuff that already exists. There might be other stuff that could benefit from that sort of thing as well.