@slaren 2023-10-07
> Are there actually already two kernels for CUDA?

CUDA has different kernels for matrix-vector and matrix-matrix multiplication. To do this, first the matrix-vector kernels would need to be updated to support matrix-matrix multiplication.