@ibehnam 2023-10-23
> Bummer. I can't figure out an universal way to determine which kernel to use when. The break-even point for the number of batches at which the matrix-matrix kernel becomes more performant than the matrix-vector kernel depends both on the specifics of the hardware which are not queryable (number of cores, memory bandwidth, FLOPs) and on the model / matrix sizes.

Optionally, we can ask the user to pass in their hardware specs. In the docs, there's already a [hint](https://github.com/ggerganov/llama.cpp/tree/master/examples/main#number-of-threads) to set `--threads` equal to the number of physical CPU cores, implying that the user can obtain this info about their system.